+++ SampleApp/Sources/UI/MainViewController.swift
@@ -243,9 +243,7 @@ private extension MainViewController {
                 return
             }
             NuguCentralManager.shared.localTTSAgent.stopLocalTTS()
-            NuguCentralManager.shared.client.asrAgent.startRecognition(initiator: initiator) { [weak self] (asrResult, _) in
-                self?.updateVoiceChrome(asrResult)
-            }
+            NuguCentralManager.shared.client.asrAgent.startRecognition(initiator: initiator)
             
             self.nuguVoiceChrome.removeFromSuperview()
             self.nuguVoiceChrome = NuguVoiceChrome(frame: CGRect(x: 0, y: self.view.frame.size.height, width: self.view.frame.size.width, height: NuguVoiceChrome.recommendedHeight + SampleApp.bottomSafeAreaHeight))
@@ -270,33 +268,6 @@ private extension MainViewController {
         }, completion: { [weak self] _ in
             self?.nuguVoiceChrome.removeFromSuperview()
         })
-    }
-    
-    private func updateVoiceChrome(_ asrResult: ASRResult) {
-        switch asrResult {
-        case .complete(let text):
-            DispatchQueue.main.async { [weak self] in
-                self?.nuguVoiceChrome.setRecognizedText(text: text)
-                ASRBeepPlayer.shared.beep(type: .success)
-            }
-        case .partial(let text):
-            DispatchQueue.main.async { [weak self] in
-                self?.nuguVoiceChrome.setRecognizedText(text: text)
-            }
-        case .error(let error):
-            DispatchQueue.main.async { [weak self] in
-                switch error {
-                case ASRError.listenFailed:
-                    ASRBeepPlayer.shared.beep(type: .fail)
-                    self?.nuguVoiceChrome.changeState(state: .speakingError)
-                case ASRError.recognizeFailed:
-                    NuguCentralManager.shared.localTTSAgent.playLocalTTS(type: .deviceGatewayRequestUnacceptable)
-                default:
-                    ASRBeepPlayer.shared.beep(type: .fail)
-                }
-            }
-        default: break
-        }
     }
 }
 
@@ -488,6 +459,34 @@ extension MainViewController: ASRAgentDelegate {
             break
         }
     }
+    
+    func asrAgentDidReceive(result: ASRResult, dialogRequestId: String) {
+        switch result {
+        case .complete(let text):
+            DispatchQueue.main.async { [weak self] in
+                self?.nuguVoiceChrome.setRecognizedText(text: text)
+                ASRBeepPlayer.shared.beep(type: .success)
+            }
+        case .partial(let text):
+            DispatchQueue.main.async { [weak self] in
+                self?.nuguVoiceChrome.setRecognizedText(text: text)
+            }
+        case .error(let error):
+            DispatchQueue.main.async { [weak self] in
+                switch error {
+                case ASRError.listenFailed:
+                    ASRBeepPlayer.shared.beep(type: .fail)
+                    self?.nuguVoiceChrome.changeState(state: .speakingError)
+                case ASRError.recognizeFailed:
+                    NuguCentralManager.shared.localTTSAgent.playLocalTTS(type: .deviceGatewayRequestUnacceptable)
+                default:
+                    ASRBeepPlayer.shared.beep(type: .fail)
+                }
+            }
+        default: break
+        }
+    }
+
 }
 
 // MARK: - TextAgentDelegate